Remove bedtime api
