+++ .env.prod
@@ -1,5 +1,5 @@
 APP_URL=https://audius.co
-BEDTIME_URL=https://di2033imt2ldz.cloudfront.net/
+EMBED_PLAYER_URL=https://di2033imt2ldz.cloudfront.net/
 API_SUBDOMAIN=api.audius.co
 ACCESS_TOKEN=lollapalooza
