+++ .env.stage
@@ -1,5 +1,5 @@
 APP_URL=https://staging.audius.co
-BEDTIME_URL=https://d3ebu1e2mbelyv.cloudfront.net/
+EMBED_PLAYER_URL=https://d3ebu1e2mbelyv.cloudfront.net/
 API_SUBDOMAIN=api.staging.audius.co
 ACCESS_TOKEN=outsidelands
