+++ src/servlets/utils/helpers.ts
@@ -201,3 +201,18 @@ export const shuffle = (a: any[]) => {
 const REDIRECT_TRACK_ID_RANGE = [416972, 418372]
 export const shouldRedirectTrack = (trackId: number) =>
   trackId >= REDIRECT_TRACK_ID_RANGE[0] && trackId <= REDIRECT_TRACK_ID_RANGE[1]
+
+
+/**
+ * Generate a short base36 hash for a given string.
+ * Used to generate short hashes for for queries and urls.
+ * @param {string} str
+ * @returns {string} hash
+ */
+export const getHash = (str: string) =>
+  Math.abs(
+    str.split('').reduce((a, b) => {
+      a = (a << 5) - a + b.charCodeAt(0)
+      return a & a
+    }, 0)
+  ).toString(36)
